requirements.txt
fastapi==0.104.1
python-multipart==0.0.6
uvicorn==0.24.0
markitdown==0.1.0
python-dotenv==1.0.0 